Add "sha1" Argument to upload file args.
Defaults 'X-Bz-Content-Sha1' to the 'sha1' argument. If this value is not
set the functionality will roll back to using the data or sha1 of the data
which is passed in. Effectively this allows callers who have already
calculated their SHA1 to send it to this function and not re-calculate it.
Prevents the unnecessary calculation from happening again.